# docs/swagger.yaml
definitions:
...
  global.ApiResponse:
    properties:
      code:
        $ref: '#/definitions/errInfos.ErrCode'
      datas: {}
      message:
        type: string
      warnings:
        items:
          $ref: '#/definitions/global.Warning'
        type: array
    type: object
...
  requests.CheckBufferRequest:
    properties:
      course_id:
        type: integer
      next_start_time:
        type: string
      prev_end_time:
        type: string
      room_id:
        type: integer
      teacher_id:
        type: integer
    required:
    - course_id
    - next_start_time
    - prev_end_time
    - room_id
    - teacher_id
    type: object
  requests.CheckOverlapRequest:
    properties:
      end_time:
        type: string
      exclude_rule_id:
        type: integer
      room_id:
        type: integer
      start_time:
        type: string
      teacher_id:
        type: integer
      weekday:
        description: 可選，如果未提供則從 StartTime 推算
        type: integer
    required:
    - end_time
    - room_id
    - start_time
    type: object
...
  requests.ReviewExceptionRequest:
    properties:
      action:
        type: string
      override_buffer:
        type: boolean
      reason:
        type: string
    required:
    - action
    type: object
...
  services.BufferCheckResult:
    properties:
      conflicts:
        items:
          $ref: '#/definitions/services.ValidationConflict'
        type: array
      valid:
        type: boolean
    type: object
...
  services.OverlapCheckResult:
    properties:
      conflicts:
        items:
          $ref: '#/definitions/services.ValidationConflict'
        type: array
      valid:
        type: boolean
    type: object
...
  services.TodaySummary:
    properties:
      changes_count:
        type: integer
      completed_sessions:
        type: integer
      has_schedule_changes:
        type: boolean
      in_progress_sessions:
        type: integer
      in_progress_teacher_names:
        items:
          type: string
        type: array
      pending_exceptions:
        type: integer
      sessions:
        items:
          $ref: '#/definitions/services.TodaySession'
        type: array
      total_sessions:
        type: integer
      upcoming_sessions:
        type: integer
    type: object
...
info:
  contact: {}
  description: API 維護文件
  title: timeLedger 阿卡莉 模板框架
  version: "1.0"
paths:
...
  /api/v1/admin/dashboard/today-summary:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/global.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/services.TodaySummary'
              type: object
      security:
      - BearerAuth: []
      summary: 取得管理員後台首頁的今日課表摘要
      tags:
      - Admin - Dashboard
...
  /api/v1/admin/scheduling/check-overlap:
    post:
      consumes:
      - application/json
      parameters:
      - description: 衝突檢查請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.CheckOverlapRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/global.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/services.OverlapCheckResult'
              type: object
      security:
      - BearerAuth: []
      summary: 檢查課程時間是否與現有排程衝突
      tags:
      - Admin - Scheduling
  /api/v1/admin/scheduling/check-room-buffer:
    post:
      consumes:
      - application/json
      parameters:
      - description: 緩衝檢查請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.CheckBufferRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/global.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/services.BufferCheckResult'
              type: object
      security:
      - BearerAuth: []
      summary: 檢查教室的緩衝時間是否足夠
      tags:
      - Admin - Scheduling
  /api/v1/admin/scheduling/check-teacher-buffer:
    post:
      consumes:
      - application/json
      parameters:
      - description: 緩衝檢查請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.CheckBufferRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/global.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/services.BufferCheckResult'
              type: object
      security:
      - BearerAuth: []
      summary: 檢查老師的緩衝時間是否足夠
      tags:
      - Admin - Scheduling
  /api/v1/admin/scheduling/exceptions:
    get:
      consumes:
      - application/json
      parameters:
      - description: 開始日期 (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: 結束日期 (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/global.ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.ScheduleException'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 取得指定日期範圍內的所有例外申請
      tags:
      - Admin - Scheduling
  /api/v1/admin/scheduling/exceptions/{id}/review:
    post:
      consumes:
      - application/json
      parameters:
      - description: 例外ID
        in: path
        name: id
        required: true
        type: integer
      - description: 審核資訊
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.ReviewExceptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/global.ApiResponse'
      security:
      - BearerAuth: []
      summary: 審核例外申請（核准/拒絕）
      tags:
      - Admin - Scheduling
  /api/v1/admin/scheduling/exceptions/all:
    get:
      consumes:
      - application/json
      parameters:
      - description: 狀態篩選：PENDING, APPROVED, REJECTED, REVOKED
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/global.ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.ScheduleException'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 取得所有例外申請（可依狀態篩選）
      tags:
      - Admin - Scheduling
  /api/v1/admin/scheduling/exceptions/pending:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/global.ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.ScheduleException'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 取得所有待審核的例外申請
      tags:
      - Admin - Scheduling
  /api/v1/admin/scheduling/expand:
    post:
      consumes:
      - application/json
      parameters:
      - description: 展開請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.ExpandRulesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/global.ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/services.ExpandedSchedule'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 展開排課規則為具體課程場次
      tags:
      - Admin - Scheduling
  /api/v1/admin/scheduling/phase-transitions:
    post:
      consumes:
      - application/json
      parameters:
      - description: 偵測請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.DetectPhaseTransitionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/global.ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/services.PhaseTransition'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 偵測課程序列中的階段轉換點
      tags:
      - Admin - Scheduling
  /api/v1/admin/scheduling/rules:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/global.ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.ScheduleRule'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 取得中心的所有排課規則
      tags:
      - Admin - Scheduling
    post:
      consumes:
      - application/json
      parameters:
      - description: 規則資訊
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.CreateRuleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/global.ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.ScheduleRule'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 建立新的排課規則
      tags:
      - Admin - Scheduling
  /api/v1/admin/scheduling/rules/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 規則ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/global.ApiResponse'
      security:
      - BearerAuth: []
      summary: 刪除排課規則
      tags:
      - Admin - Scheduling
    put:
      consumes:
      - application/json
      parameters:
      - description: 規則ID
        in: path
        name: id
        required: true
        type: integer
      - description: 規則資訊
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.UpdateRuleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/global.ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.ScheduleRule'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 更新排課規則
      tags:
      - Admin - Scheduling
  /api/v1/admin/scheduling/rules/{id}/exceptions:
    get:
      consumes:
      - application/json
      parameters:
      - description: 規則ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/global.ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.ScheduleException'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 取得指定規則的所有例外申請
      tags:
      - Admin - Scheduling
  /api/v1/admin/scheduling/rules/check-lock:
    post:
      consumes:
      - application/json
      parameters:
      - description: 檢查請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.CheckRuleLockStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/global.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/services.RuleLockStatus'
              type: object
      security:
      - BearerAuth: []
      summary: 檢查規則是否已超過異動截止日
      tags:
      - Admin - Scheduling
  /api/v1/admin/scheduling/validate:
    post:
      consumes:
      - application/json
      parameters:
      - description: 完整驗證請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.ValidateFullRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/global.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/services.FullValidationResult'
              type: object
      security:
      - BearerAuth: []
      summary: 完整驗證排課（硬衝突 + 緩衝檢查）
      tags:
      - Admin - Scheduling
...
  /api/v1/teacher/me/exceptions:
    post:
      consumes:
      - application/json
      parameters:
      - description: 例外資訊
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.CreateExceptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/global.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ScheduleException'
              type: object
      security:
      - BearerAuth: []
      summary: 建立新的例外申請
      tags:
      - Teacher - Scheduling
...
schemes:
- http
swagger: "2.0"